Problem: upgrading Omnigres container images
Once you upgrade the image, previous versions of Omnigres extensions are gone.
And if you had them provisioned in your database, well, good luck.

Solution: preserve previous extensions (effectively .control, .sql and .so files)

We do this by building an image off of the latest released image and copy pkglibdir
and sharedir/extension into the new image, before installing Omnigres extensions.

This allows us to specify new default versions and override whatever files we'd
like to override.

Since all releases go through a single point of CI in master, we'll release them
sequentially. It may get a little dicey when builds fail and we don't retry them
in time, but even then it's probably okay because it's still going to pick up the latest
version. As long as we don't retry them out of order (can we prevent that?)