Problem: A few edge cases are not covered by the current implementation: * null flag argument * nested begins * unfinished transaction * using start transaction syntax
Solution: Regarding the start transaction syntax and the nested begins
we have to deparse the T_TransactionStmt to properly identify cases of
starting and ending transactions. This will also solve other edge cases
not described above such "end" commands before a commit.

We identify start and end comparing the deparsed to COMMIT/ROLLBACK
since the START TRANSACTION would be more nuanced and harder to compare.

The null argument check was added to the top of the function as well as
an if after the loop to flush any lingering transactions.