Problem: changes in upcoming Postgres
We learn about them too late, when the release is arriving. This means
we are late to provide timely feedback, and it takes time to integrate
new versions in bulk.

Solution: track master on every CI build

We will not fail the job if it fails, but we'll be aware of failures.

To accomplish this, I extended the cmake definition to allow specifying
sha1 and `master` references in `PGVER`

This change also addresses issues that were spotted during the upgrade.

Most notably:

* SQL AST changes
* `ExecutorRun` dropping the `excecute_once` parameter