Problem: Postgres' information schemas are not great
The information is there but it is difficult to consume and the lack
referential clarity.

Solution: use Aquameta's `meta` extension

Eric Hanson from Aquameta has done an amazing, tedious job of putting
it all together.

Let's see if this can become an important part / dependency of omni_schema!

I've fitted it into instantiate template to avoid hard-coding the schema.