Problem: analyzing differences between schemas
The task can be a bit daunting.

Solution: extract added/removed/changed views
to provide high level of diffing.

This is done by creating a schema with views that connect two other schemas.
This is not ideal (would it be interesting to be able to change the target
schemas based on some configuration parameter?) but works for now.

Supplementary to this, added the following abilities:

* Materializing meta schema (useful for baseline comparison)
* Accessing meta schema over FDW (useful for inter-database comparisons)