Problem: diffing more complicated changes
Now that we've done some basic table stuff with a fairly small surface, we need
to start covering other meta-information.

However, it becomes less trivial as we need to be able to identify changes
without comparing rows.

Solution: go 6NF (sixths normal form) on the meta design

The reason for this is that with 6NF we can identify change of every particular
fact very easily without doing row by row comparison.

To this effect, this removes baseline/changed views as added/changed is all you
need.

This changes an underlying assumption on the integrated `meta` code, and it'll
continue diverging from it as we change it. Since we've already had tables, I
had to go back and make sure the design aligns.

I've also enabled pg_catalog types to be included in the `type` view for
referential integrity.