Problem: looking up next available PR number
When preparing a pull request, in order to add a changelog line,
I need a _future_ PR number to reference. This forces me to go to the
repository to check what was the latest issue or PR.

Solution: write a script to automate this

The script requires GitHub CLI tooling to be installed and configured.