Problem: sending HTTP response before committing the transaction
This may present a timing problem when we want to query recently modified data
based on the fact that the response has already been sent.

We've observed this in sporadic test failures.

Solution: ensure we commit first

This is not yet perfect case coverage as this doesn't include WebSocket sending
as it will send messages eager before transanction is committed.

This generally raises a good question: should we apply the same transactional
guarantees to outbound WebSocket messages? Should we provide an option for this,
defaulting to either choice?