Problem: Postgres 13 is too old
We keep running into its limitations. It stops being supported in Nov 2025,
so in about 9 months.

Solution: drop supporting it now

This will redirect energy required to support it to something more useful.

Note that this doesn't change some of the syntax or conditional compilation
that pertains to Postgres 13. We just stop building and testing on it. Soon
enough, some extensions will stop working on it.