Problem: no response from false websocket_handler
If `omni_httpd.websocket_handler()` returns false, no HTTP response is
sent. And it seems like the worker just hangs for some reason -- it
never responds to future requests -- even if the connection is closed by
the peer.

However, if the handler returns null, it does send an HTTP response that
does not upgrade the connection to a WebSocket.

But the comment for the handler indicates that any non-true value should
be treated the same:

    If it returns anything but true, the connection is rejected.

Solution: send a response in any case

This change should send an HTTP even if false is returned and only
upgrade if the handler returns true.

I'm still not sure why the worker seemed to not receive further
requests when it failed to send a response here. (Though omni_httpd.stop
and start would restart them.) Even when the connection was closed by
the peer, I don't think the worker was closing its file descriptor for
the socket. But this might have something to do with how the h2o library
works that I'm not familiar with.