Problem: sum types won't compare composite types
It would compare other types just fine as evidenced by tests. But if we create
a composite type, put it into a sum type, and try to compare, it won't find it.

Solution: ensure we use operator selection techniques from Postgres

This ensures we select a right candidate even if there is no direct match.