Problem: impossible to track dependencies in schemas
This can be very useful to identify which parts of schema are brought in by the
application and which ones are coming from, say, extensions.

Solution: introduce omni_schema.dependency view

It's not implemented very performantly for now but it does the job. The
coverage is not exhaustive yet.

I've also added `language_id` type since dependencies can happen on language
and this is visible in a barebones, template database.