Problem: sum type equality comparison is slow
We keep looking up the variant and the variant comparison functions a lot.

Solution: cache these

Bound to a xmin, we keep a cache of what we know about the types. This is still
imperfect because we don't preserve the cache across the xmin boundary where we
could have. For that, we would to need to implement a relcache handler in
omni_types to refresh the cache whenever the underlying table changes.

Also, the cache is only used in variant lookup but not in in/out functions.