Problem: omni_httpd leaks file descriptors
WebSockets cause omni_httpd to leak file descriptors, mainly through the
unix socket proxy tunnel thingy and but also not closing the TCP
connection. Eventually, it runs out of file descriptors and cannot
accept new connections. It can be particularly troublesome if accept()
fails with "Too many open files" and gets stuck trying again and again
failing each time indefinitely. I think better error handling and
reporting would valuable in future work.

Solution: close file descriptors for opened sockets

1. `on_ws_message` might not call h2o_websocket_close on the TCP socket
   for a WebSocket

   `on_ws_message` would see `arg->opcode == WSLAY_CONNECTION_CLOSE` and
   `free(conn->data)` and assign it to NULL.

   Then `on_ws_message` would see `arg == NULL` but not enter the branch
   to call `h2o_websocket_close()` because data is NULL at that point.

   This only frees the data and closes the socket when when
   `arg == NULL`.

2. When a postgres connection process calls `websocket_send` from a
   SQL query, it needs to forward the message to whatever httpd worker
   holds the underlying connection for that WebSocket.

   In doing so it opens a socket but never closes it, it stores it in a
   collection to re-use it later but it does eventually leak and long
   running connection processes can run into problems for this reason.

   This change simply closes the socket each time and removes the
   collection where open sockets were stored. This does have negative
   performance impact in terms of speed but positively impacts
   performance as the process is not doomed to run out of file
   descriptors. :)

   The two `send()` calls have been replaced with one `sendmsg()` using
   scatter-gather but I'm not sure this helps with speed all that much.
   (There is a lot of noise when I try to measure this extension on my
   machine.)

   So there is certainly room for a more efficient design, maybe one
   that includes the collection but is length limited and closes least
   recent sockets.

3. Also in `on_ws_message`, the unix listen socket for accepting
   connections from `websocket_send` was not being closed. The unix
   socket is opened when the WebSocket is upgraded. With this change,
   the file descriptor is copied to a `on_ws_message_data_t` structure
   pointed to by the h2o connection where it can be closed by
   `on_ws_message` when the WebSocket is closed.

Finally, h2o provides uuids for connections through
`const char *h2o_conn_get_uuid(h2o_conn_t *)`. It may be simpler to use
that and remove the uuid generation and formatting code in this
extension. As far as I know, once an HTTP connection is upgraded to a
WebSocket, it cannot be used for any other HTTP requests. Even in HTTP2
which supports streaming, once it's upgraded, it stops being whatever it
was and the connection is a WebSocket from then on. So from what I
understand, it's not as if more than one WebSocket can run on one TCP
connection; it can't be upgraded again in a separate stream. So if the
WebSocket uuid is just the connection uuid, it would still remain unique
across WebSockets.