Problem: omni_schema meta doesn't capture all priviliges
It does capture table privileges but doesn't go beyond that. However, capturing
correct set of grants is absolutely essential for enforcing any kind of security policy.

Solution: introduce a unified `acl` view

It doesn't look very normalized – it has to do with the lack of a primary key
and atomicity of grants.

The most questionable decision so far is including the `default` column. It indicates
if this was an actual grant or this is how it looks by default. I've included it so that
`acl` view faithfully represents the state of possibilities.