Problem: use of uuidv7() in omni_id without search_path
Since `uuidv7` is most often accessed through omni_polyfill, we often need to
set the whole environment to have polyfill available in the search path just for
the `nextval` to work.

This is hugely inconvenient and confusing.

Solution: simplify the code to allow for the reference to uuidv7 to be used