Problem: generializing omni_credentials
When we want to use them across different subsystems and extensions, they lack
specificity. It's just an encrypted key/value store.

It's not clear what naming convention should be used, for example.

Solution: expand the concept to include further metadata

"What kind of credential?" – `kind`
"Who can have access to it?" – `principal`
"What is the scope of the credential?" – `scope`

This will allow much better "notation". Since we're getting all these details,
it now makes sense to enable RLS on this table and by default, filter by principal.
Users can add more policies to it. The reason why this metadata is not factored out
is so that RLS _can_ be used there without sub-queries.

This removes the `credentials_master_password` function that was effectively leaking the
encryption key.