Problem: inability to run workloads in other backends
Sometimes we need to execute statements in other backends, sparing the current
one. In more complicated cases, we need to run arbitrary [compiled] code in a
pool of workers, like an HTTP handler.

omni_httpd solves this by having its own pool of workers. But it is not usable
for anything else.

Solution: provide a generalized worker pool

For now, it's pretty simplistic – just runs a pre-set number of workers that
would consume scheduled messages as they come.