Problem: can't use `security definer` procedures in omni_httpd
Problems start appearing when we try to use multi-transactional
capabilities (the very reason we support procedures) as the case
of security-definer procedures is not currently supported in Postgres.

Solution: introduce `handler_role` relation

This relation is external annotation of desires roles, as opposed to
security definer.

We could have set this from security-definer procedure's owner
information, but then `fmgr_security_definer` would kick in and modify
the security context, which `StartTransaction` does not expect and
will fail. So we're forced to externalize this.