Problem: after-handler middleware in omni_httpd
If we need to change the response in any way after the handler was
called, we can't really do much. This can be very useful for things like
CORS handlers or anything that changes the response.

Solution: only stop processing "handlers" after one is executed
but continue with the rest of middleware handlers

(We really should rename them to something more sensible – middleware
"after" doesn't make sense logically)