Problem: loading extensions in presence of live backends
If there are two backends, A and B and then A installs an extension, it will
not always be initialized (`_Omni_init`) in B.

This is a fairly critical bug.

Solution: ensure we track respective invalidations

In Postgres 18, EXTENSIONOID is a syscache. Prior to that, we'll use PROCOID
syscache as a proxy with a workaround for extensions that export no functions.