Problem: account categorization balancing failures
They occur in transactions that are not supposed to fail.

Solution: ensure we update balances first

If this doesn't happen, we have problems with:

1) actual reflected balances available to other triggers
2) interactions with the cache and our current marking for "touched"
   accounts.

Our current system of account caches and computation of balances is fairly
fragile and would benefit from a re-design (and a port to Cppgres).

This solution is a bit fragile as this depends on the naming-based order, but
this fixes the problem for now before the next major revision is ready.