Problem: versions.txt can potentially mislist
What if we forget an extension? This just happened – omni_email went
unversioned.

What if we update an extension but forget to bump the version up?

Solution: move VERSION to add_postgresql_extension
and sync it back to versions.txt

This still doesn't solve the problem of forgetting to bump up, but this is a
start.