Add omni_northwind extension
This extension installs the legendary Northwind database, that
was used as a demo database in Microsoft Access and later also
in SQL Server.

The idea is to use it as a general test dataset that other extensions
can depend on. Other extensions can also extend this data with
more modern features. For example an AI extension would add pgvector
data, and so on.