Problem: watching Kubernetes resources
Right now the only operation possible is to list resources. It may be
generally sufficient, but sometimes watching for event is more
preferrable.

Solution: introduce a limited watching capability

It's not really "streaming" as it forces a timeout to finish and return
the results (until we figure out a better model for it)

This also pushed me to make it possible to make multiple API requests at
the same time to be able to watch on multiple resource collections at once.