Problem: insert into and updating omni_kube views
We don't see updated resources if we use `returning` and this makes it much
more difficult to figure things out, like what's the generated name of the
resource? Or what's the uid?

Solution: ensure we grab the newly updated object

This will allow us to do `update ... returning` and `insert ... returning`
with the new data.