Problem: developing Kubernetes operators/controllers
This is still fairly difficult. Resource views in omni_kube are helpful
but they don't help managing state – they just represent what's there
now.

Solution: add a concept of resource tables

These act as replicas and exhibit many properties of resource views
(they are naturally updatable and the updates propagate back to
Kubernetes), but they maintain a stable view that can be refreshed
manually (or scheduled).

What's even more interesting is that they can be part of trigger
cascades, enabling more logic to be embededded.

They are, however, not respecting the boundaries of transactions and
changes are reflected in Kubernetes immediately.