Problem: omni_kube resource tables aren't transactional
While it is not a huge issue, it is not perfect either – in some cases
it is desirable to at least imitate some sort of transactional
boundaries.

Solution: simulate read-committed isolation
by dry-running changes and pooling them up until commit time.

I've removed API per-statement caching. It hit a newly discovered bug in
omni_var: per-statement variables don't work well in deferred constraint
triggers because there's no portal anymore. Fix pending, but in the
meantime, it allowed me to reflect on whether we actually need this
functionality in omni_kube. It was there to establish consistency in
joined views that stem from the same HTTP request. Is this that much of
a deal now where we just represent resources? Current answer: no. Hence,
let's remove this.