Problem: omni_kube permission to modify omni_kube.refresh
Solution: make `omni_kube` an omni extenion

This allows us to always define the variable, because otherwise Postgres
trips up on attempting to set a config variable in the function
definition.

My minimal reproduction case (without an extension) is:

```sql
create role r;
grant create on schema public to r;
set role r;
set a.b = 2; -- works fine
create function f() returns bool set a.b = 1 return true; -- FAILS
--- permission denied to set parameter "a.b"
```

I've traced this error in the debugger, it's `CreateFunction` ->
`GUCArrayAdd` -> `validate_option_array_item`

Particularly, it is throwing the error after doing
`pg_parameter_aclcheck(name, GetUserId(), ACL_SET) == ACLCHECK_OK`,  as
it is shown as a placeholder.

So, basically, while we're allowed to `set a.b = c` as a command, this
only creates a placeholder and that does not pass the validation
[indirectly] imposed by `CreateFunction`.

NB: While at it, figured that load_kubeconfig should be security
definer so we don't have to give server file reading roles to the user.