Problem: shutting down database with omni_worker running
Even when we drop the extension, it still keeps running and results in
blocking of the database, manifested by:

```
database "<name>" is being accessed by other users
```

Solution: terminate on extension drop
and provide explicit stop similar to omni_httpd

To be frank, current implementation is not great and has been produced
under timing pressure – the round-robin of passing with exit request
with an eventual cleanup neither pretty nor efficient. In the name of
resolving a pressing concern, this is a somewhat acceptable solution for
now.